% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_one_step_test_typeDS.R
\name{grad_one_step_test_typeDS}
\alias{grad_one_step_test_typeDS}
\title{aggregate function}
\usage{
grad_one_step_test_typeDS(
  data,
  current_parameters,
  aggr_grad,
  covs,
  bins,
  conts,
  visits,
  labels,
  n_states,
  k_tup_initial,
  k_tup_trans,
  k_tup_em,
  sim_no,
  comments
)
}
\arguments{
\item{data}{}

\item{current_parameters}{a vector of current parameters}

\item{aggr_grad}{a vector of aggregated gradients}
}
\value{
a list containing:
}
\description{
aggregate function
}
