% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_one_stepDS.R
\name{grad_one_stepDS}
\alias{grad_one_stepDS}
\title{aggregate function}
\usage{
grad_one_stepDS(
  data,
  current_parameters,
  aggr_grad,
  covs,
  bins,
  conts,
  visits,
  labels,
  n_states,
  k_tup_initial,
  k_tup_trans,
  k_tup_em,
  sim_no = "fed_test",
  comments = ""
)
}
\arguments{
\item{current_parameters}{a vector of current parameters}

\item{aggr_grad}{a vector of aggregated gradients}

\item{model_params}{a list of model parameters including:}
}
\value{
a list containing:
}
\description{
aggregate function
}
