% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_model_paramsDS.R
\name{return_model_paramsDS}
\alias{return_model_paramsDS}
\title{aggregate function}
\usage{
return_model_paramsDS(
  data,
  current_parameters,
  aggr_grad,
  covs,
  bins,
  conts,
  visits,
  labels,
  n_states,
  k_tup_initial,
  k_tup_trans,
  k_tup_em,
  sim_no,
  comments
)
}
\arguments{
\item{data}{}

\item{current_parameters}{a vector of current parameters}

\item{aggr_grad}{a vector of aggregated gradients}
}
\value{
a list containing:
}
\description{
aggregate function
}
