% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_only_model_paramsDS.R
\name{test_only_model_paramsDS}
\alias{test_only_model_paramsDS}
\title{aggregate function}
\usage{
test_only_model_paramsDS(data, current_parameters, aggr_grad)
}
\arguments{
\item{data}{}

\item{current_parameters}{a vector of current parameters}

\item{aggr_grad}{a vector of aggregated gradients}
}
\value{
a list containing:
}
\description{
aggregate function
}
