% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_one_stepDS.R
\name{grad_one_stepDS}
\alias{grad_one_stepDS}
\title{aggregate function}
\usage{
grad_one_stepDS(data, current_parameters, aggr_grad, model_params)
}
\arguments{
\item{current_parameters}{a vector of current parameters}

\item{aggr_grad}{a vector of aggregated gradients}

\item{model_params}{a list of model parameters including:}
}
\value{
a list containing:
}
\description{
aggregate function
}
